Expose internal details from composition module
Fixes #48
